Benchmark regions

Different world regions described in terms of regions and countries. Data contains lists of vectors, vectors represented in pairs of Long Names (e.g. Country name) and ISO codes (Country ISO 2-char code)

data.benchmark.regions

Format

a list of following pairs of vectors

BenchRegions, BenchRegions_ISO_colnames

Set of benchmark regions

Global, Global_ISO

Global region

Africa, Afrika_ISO

Africa

Brazil, Brazil_ISO

Brazil

China, China_ISO

China

EEurope_Eurasia, EEurope_Eurasia_ISO

Eastern Europe and Eurasia

EU, EU_ISO

Europe

India, India_ISO

India

Japan, Japan_ISO

Japan

LatinAmerica, LatinAmerica_ISO

Latin America

MiddleEast, MiddleEast_ISO

Middle East

NonOECD, NonOECD_ISO

non-OECD countries

NonOECDAsia, NonOECDAsia_ISO

non-OECD countries of Asia

OECD, OECD_ISO

OECD countries

OECDAmericas, OECDAmericas_ISO

OECD countries of Americas

OECDAsiaOceania, OECDAsiaOceania_ISO

OECD countries of Asia and Oceania

OECDEurope, OECDEurope_ISO

OECD countries of Europe

Russia, Russia_ISO

Russia

SouthAfrica, SouthAfrica_ISO

South Africa

US, US_ISO

United States of America

OECDAsiaOceaniaWoJP, OECDAsiaOceaniaWoJP_ISO

OECD countries of Asia and Oceania wihout Japan

OECDAmericasWoUS, OECDAmericasWoUS_ISO

OECD countries of Americas without US

LatinAmericaWoBR, LatinAmericaWoBR_ISO

Latin America without Brazil

AfricaWoZA, AfricaWoZA_ISO

Africa without South Africa

EEurope_EurasiaWoRU, EEurope_EurasiaWoRU_ISO

Eastern Europe and Eurasia without Russia

NonOECDAsiaRest, NonOECDAsiaRest_ISO

non-OECD countries of Asia - the rest

NonOECDRest, NonOECDRest_ISO

non-OECD countries - the rest