Match names using blocking

split names and dictionary values into blocks and make matching within corresponding blocks

matchNamesByBlocks(names, nblocks, dictionary, dblocks, threshold = 0.95,
  return = "closest")

Arguments

names

name to compare against dictionary

nblocks

vector that specifies blocks for names; should have the same length as names

dictionary

set of plausible values to chose from

dblocks

vector that specifies blocks for dictionary elements; should have the same length as names

threshold

threshold value, minimum score to filter out results with

return

what values to return from c("closest", "all"). Return type "closest" works much faster and returns only 1 match per name with highest score.

Value

dataframe of structure described as result of matchNames function, extended with Segment column that represents block.