Prepare loanbook data for matching

function that prepares local loanbook data for matching. It takes loanbook and override rules and produces dataframe of universal structure for matching.

process.prepLBK4match(loanbook, selections, override)

Arguments

loanbook

local loanbook with minimum set of required fields (see loanbook.example)

selections

list of named selections from loanbook, each selection in list must contain following attributes: level (level of mathcing); ids (Column name with IDs); values (Column name with values used for mathcing). Besides this, selection may contain following optional attributes: value.type (type of values provided) simplify (flag that runs name simplification of values provided - see simplifyName function for details) cut.ownership (flag that instructs simplify algorithm to cut ownership type)

override

local override rules for matching: dataframe

Value

data.frame with column set: CustomerID, Level, Relation, ValueType, OriginalValue, MatchValue, SectorClassification